requirements.txt
streamlit
httpx
python-dotenv
fastapi
uvicorn
pydantic
